+++ app/javascript/__tests__/articles/ArticleDetail.test.tsx
@@ -1,29 +1,13 @@
 import "@testing-library/jest-dom/extend-expect";
 import { cleanup, render } from "@testing-library/react";
+import { createLocation, createMemoryHistory } from "history";
 import * as React from "react";
-import APIConsumer from "../../common/api-consumer";
-import { Article, ArticleDetailMatchParams, User } from "../../common/types";
-import ArticleDetail from "../../components/articles/ArticleDetail";
 import { match } from "react-router";
-import { createMemoryHistory, createLocation } from "history";
+import { ArticleDetailMatchParams } from "../../common/types";
+import ArticleDetail from "../../components/articles/ArticleDetail";
+import { mockArticle, consumer } from "./__helpers__/test-data";
 
 const history = createMemoryHistory();
-
-const consumer: APIConsumer = new APIConsumer();
-
-const mockUser: User = {
-  firstName: "Test",
-  lastName: "User",
-  id: "1",
-};
-
-const mockArticle: Article = {
-  id: "10",
-  title: "Some Fake Title",
-  content: "Some dummy content",
-  createdAt: "2020-06-24T16:10:53.448Z",
-  user: mockUser,
-};
 
 const createRouteComponentProps = (): ArticleDetailMatchParams => {
   const path = `/articles/:id`;
